Completed to do
Everything is working as it should.

Order of operations:
1. Set global myTable variable
2. Create the table in a for loop, populate data, stripe table
3.  Total contributors
4. Total contribution amount, for loop to iterate through specific cell
in each row and add amounts then format the output

### follow up
parseInt, parseFloat kept throwing NaN in the total contributions amount
for loop. The .substring(1) is used to get rid of the $ sign formatting
when we created the table. Is this an acceptable solution? Should this
be a regEx? The output matches finalpage.jpg, however I want to be sure
that this solution is correct, or if not how to implement the correct
solution.